split XmlFactoryWriterTests.cs into smaller test classes and comment out all but one
