+++ src/System.Private.Xml/tests/Writers/XmlWriterApi/System.Xml.RW.XmlWriterApi.Tests.csproj
@@ -18,9 +18,23 @@
     <Compile Include="ErrorCondition.cs" />
     <Compile Include="NamespaceHandlingTests.cs" />
     <Compile Include="ReaderParamTestCase.cs" />-->
-    <Compile Include="TestExtensions.cs" />
+    <!--<Compile Include="TCAutoCL.cs" />
+    <Compile Include="TCCheckChars.cs" />
+    <Compile Include="TCCloseOutput.cs" />
+    <Compile Include="TCDefaultWriterSettingsTests.cs" />
+    <Compile Include="TCFlushClose.cs" />
+    <Compile Include="TCFragmentCL.cs" />
+    <Compile Include="TCIndent.cs" />
+    <Compile Include="TCIndentChars.cs" />
+    <Compile Include="TCNewLineChars.cs" />
+    <Compile Include="TCNewLineHandling.cs" />
+    <Compile Include="TCNewLineOnAttributes.cs" />
+    <Compile Include="TCOmitXmlDecl.cs" />
+    <Compile Include="TCStandAlone.cs" />
+    <Compile Include="TCWriterSettingsMisc.cs" />
+    <Compile Include="TCWriterWithMemoryStream.cs" />
+    <Compile Include="TestExtensions.cs" />-->
     <Compile Include="WriteEndDocumentOnCloseTest.cs" />
-    <Compile Include="XmlFactoryWriterTests.cs" />
     <!--<Compile Include="XmlWriterTestCaseBase.cs" />
     <Compile Include="XmlWriterTestModule.cs" />-->
   </ItemGroup>
